+++ src/objects/MovingGameObject.java
@@ -2,25 +2,25 @@ package objects;
 
 public class MovingGameObject extends GameObject {
 	
-	private int hSpeed, vSpeed;
+	private Double hSpeed, vSpeed;
 
-	public int gethSpeed() {
+	public Double gethSpeed() {
 		return hSpeed;
 	}
 
-	public void sethSpeed(int hSpeed) {
+	public void sethSpeed(Double hSpeed) {
 		this.hSpeed = hSpeed;
 	}
 
-	public int getvSpeed() {
+	public Double getvSpeed() {
 		return vSpeed;
 	}
 
-	public void setvSpeed(int vSpeed) {
+	public void setvSpeed(Double vSpeed) {
 		this.vSpeed = vSpeed;
 	}
 
-	public MovingGameObject(int hPos, int vPos, boolean collision) {
+	public MovingGameObject(Double hPos, Double vPos, boolean collision) {
 		super(hPos, vPos, collision);
 	}
