+++ src/objects/StaticGameObject.java
@@ -2,7 +2,7 @@ package objects;
 
 public class StaticGameObject extends GameObject{
 
-	public StaticGameObject(int hPos, int vPos, boolean collision) {
+	public StaticGameObject(Double hPos, Double vPos, boolean collision) {
 		super(hPos, vPos, collision);
 	}
